// komg/komgo-member-api-credit-lines/src/business-layer/notifications/enums.ts
export enum PRODUCT_ID {
  TradeFinance = 'tradeFinance'
}

export enum NOTIFICATION_TYPE {
  CLInfo = 'CL.info',
  DepositLoanInfo = 'CL.DepositLoan.info',
  CLRequest = 'CL.CLRequest'
}

export enum NOTIFICATION_LEVEL {
  info = 'info',
  warning = 'warning',
  danger = 'danger'
}

export enum SUB_PRODUCT_ID {
  RiskCover = 'rd',
  BankLine = 'lc'
}